TDM: Add precipitation adjustment factor
precipitation rates extracted form CMIP5 seem to be a factor of 10 off, despite all documentation having the correct units.  I'm sure it's wrong because scalign it up to annual preciptiation is implausibly low (137 mm/yr instead of ~1000 mm/yr for Chicago) and all the trees die when I try to run ed.  :angry: