potential speedup: only generate 1 set of values
Rather than generate an arbitrary number of values to try and pick from, we can just generate 1 random set of values at a time to propagate uncertainty.  If you wanted to use the "mean" option to not really generate more uncertainty, then just use the mean model betas in the same way we do with a single enesmble member.  This should reduce memory needs and possibly computational time.  However, there are now a lot of artifacts int he code that could probably be cleaned over.  I don't *think* we can just get rid of the ensemble loop since we train the bias-correction on each enesmble member that's read in, but I'm not positive about that.